+++ calva-fmt/extension.ts
@@ -21,11 +21,10 @@ function activate(context: vscode.ExtensionContext) {
     vscode.languages.setLanguageConfiguration("clojure", ClojureLanguageConfiguration);
     context.subscriptions.push(vscode.commands.registerTextEditorCommand('calva-fmt.formatCurrentForm', formatter.formatPositionCommand));
     context.subscriptions.push(vscode.commands.registerTextEditorCommand('calva-fmt.inferParens', inferer.inferParensCommand));
-    context.subscriptions.push(vscode.commands.registerTextEditorCommand('calva-fmt.tabIndent', inferer.tabIndentCommand));
-    context.subscriptions.push(vscode.languages.registerOnTypeFormattingEditProvider("clojure", new FormaOnTypeEditProvider,
-        "\n"));
+    context.subscriptions.push(vscode.commands.registerTextEditorCommand('calva-fmt.tabIndent', (e) => { inferer.tabIndentCommand(e, " ", true) }));
+    context.subscriptions.push(vscode.commands.registerTextEditorCommand('calva-fmt.tabDedent', (e) => { inferer.tabIndentCommand(e, " ", false) }));
+    context.subscriptions.push(vscode.languages.registerOnTypeFormattingEditProvider("clojure", new FormaOnTypeEditProvider, "\n"));
     context.subscriptions.push(vscode.languages.registerDocumentRangeFormattingEditProvider("clojure", new RangeEditProvider));
-
     const api = {
         formatPosition: formatter.formatPosition,
         formatRange: formatter.formatRange,